Moving translation to client component and keeping page as server component
