--- src/app/[locale]/dashboard/@board/layout.tsx
@@ -1,11 +1,0 @@
-import { NextIntlClientProvider, useMessages } from 'next-intl';
-import pick from 'lodash.pick';
-
-export default function BoardPage({ children }: { children: React.ReactNode }) {
-  const messages = useMessages();
-  return (
-    <NextIntlClientProvider messages={pick(messages, 'Board')}>
-      {children}
-    </NextIntlClientProvider>
-  );
-}
